backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic[email]==2.11.3
python-dotenv==1.1.0
httpx==0.28.1
sqlalchemy[asyncio]==2.0.49
asyncpg==0.31.0
alembic==1.18.4
greenlet==3.4.0
slowapi==0.1.9
PyJWT==2.10.1
python-multipart==0.0.20
anthropic==0.40.0
bcrypt==4.3.0
stripe==11.4.1
